(pub) improving the tickets rendering for email attachments (there is still a small bug running, but nothing major) (tck) importing the good way to render tickets from pub for transactions (liPassbookPlugin) adding a condition, that removes a PHP_WARNING

+++ plugins/liPassbookPlugin/lib/liPassbook.class.php
@@ -260,6 +260,8 @@ class liPassbook
   protected function rmPkpass($path = NULL)
   {
     $path = $path ? $path : $this->getPkpassPath();
+    if ( !file_exists($path) )
+      return $this;
     if ( !is_dir($path) )
     {
       unlink($path);
